Deploy Soroban smart contract to Stellar testnet
 Contract successfully deployed
- Contract ID: CB7K36JL25A4VO67X6FLIKCCGA2YJWI35CXZZGZ6FBV3UOQX6LLSPN5Q
- Deployer: GD4DKHCWBTQM6TMZHZT75WGC5OWMSKQETGUL42AEJGNBUYYTRZBJWEPT
- Network: Stellar Testnet
- WASM Size: 2,742 bytes

 Added comprehensive contract documentation
- CONTRACT_INFO.md with all deployment details
- Transaction hashes and explorer links
- Integration instructions for backend
- Updated README with contract information

 Configuration updates
- Added .env with contract ID
- Saved contract ID to .contract-id-testnet
- Fixed type casting issues in contract code

 Explorer Links:
- https://stellar.expert/explorer/testnet/contract/CB7K36JL25A4VO67X6FLIKCCGA2YJWI35CXZZGZ6FBV3UOQX6LLSPN5Q
- https://lab.stellar.org/r/testnet/contract/CB7K36JL25A4VO67X6FLIKCCGA2YJWI35CXZZGZ6FBV3UOQX6LLSPN5Q

+++ contracts/src/lib.rs
@@ -23,7 +23,7 @@ impl SubscriptionContract {
     pub fn subscribe(env: Env, subscriber: Address) -> Subscription {
         subscriber.require_auth();
 
-        let current_ledger = env.ledger().sequence();
+        let current_ledger = env.ledger().sequence() as u64;
         let expires_at = current_ledger + 518_400; // ~30 days
 
         let subscription = Subscription {
@@ -45,7 +45,7 @@ impl SubscriptionContract {
         let key = (SUBSCRIPTION, subscriber);
         
         if let Some(sub) = env.storage().persistent().get::<_, Subscription>(&key) {
-            let current_ledger = env.ledger().sequence();
+            let current_ledger = env.ledger().sequence() as u64;
             sub.expires_at > current_ledger && sub.scans_remaining > 0
         } else {
             false
@@ -63,7 +63,7 @@ impl SubscriptionContract {
             .get::<_, Subscription>(&key)
             .expect("No subscription found");
 
-        let current_ledger = env.ledger().sequence();
+        let current_ledger = env.ledger().sequence() as u64;
         assert!(sub.expires_at > current_ledger, "Subscription expired");
         assert!(sub.scans_remaining > 0, "No scans remaining");
 
@@ -87,7 +87,7 @@ impl SubscriptionContract {
         subscriber.require_auth();
 
         let key = (SUBSCRIPTION, subscriber.clone());
-        let current_ledger = env.ledger().sequence();
+        let current_ledger = env.ledger().sequence() as u64;
 
         let mut sub = if let Some(existing) = env.storage().persistent().get::<_, Subscription>(&key) {
             // Extend existing subscription
